+++ NEWS.md
@@ -3,6 +3,7 @@
 ## 0.0.2
 
 - Deprecate `com_serve`, it's no longer needed.
+- Rewritten in Typescript.
 
 ## 0.0.1